Refactoring graph splitter and fix minor bugs.
