+++ federatedscope/core/splitters/generic/__init__.py
@@ -2,6 +2,6 @@ from __future__ import absolute_import
 from __future__ import print_function
 from __future__ import division
 
-from federatedscope.core.splitters.splitter_builder import get_splitter
+from federatedscope.core.splitters.generic.lda_splitter import LDASplitter
 
-__all__ = ['get_splitter']
+__all__ = ['LDASplitter']
